+++ Riot/Categories/MXKRoomBubbleTableViewCell+Riot.h
@@ -16,6 +16,8 @@
 
 #import <MatrixKit/MatrixKit.h>
 
+@class CircleProgressView;
+
 /**
  Action identifier used when the user pressed edit button displayed in front of a selected event.
  
@@ -148,6 +150,8 @@ extern NSString *const kMXKRoomBubbleCellCallBackButtonPressed;
  */
 + (CGFloat)attachmentBubbleCellHeightForCellData:(MXKCellData *)cellData withMaximumWidth:(CGFloat)maxWidth;
 
+- (IBAction)onProgressLongPressGesture:(UILongPressGestureRecognizer*)recognizer;
+
 /**
  Blur the view by adding a transparent overlay. Default is NO.
  */
@@ -162,5 +166,10 @@ extern NSString *const kMXKRoomBubbleCellCallBackButtonPressed;
  The marker view displayed in front of the marked component (if any).
  */
 @property (nonatomic) UIView *markerView;
+
+/**
+ Message tick views (sending, sent) displayed alongside the related component.
+ */
+@property (nonatomic) NSArray *messageStatusViews;
 
 @end