+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changes to be released in next version
 =================================================
 
 ✨ Features
- * 
+ * Improve the status of send messages (sending, sent, received, failed) (#4014)
 
 🙌 Improvements
  *